chore: refactor the asset Redux code into multiple files
