+++ package.json
@@ -16,6 +16,7 @@
     "@babel/register": "^7.25.9",
     "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.14.0",
+    "@reduxjs/toolkit": "^2.6.1",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
@@ -25,8 +26,8 @@
     "@types/jsdom": "^21.1.7",
     "@types/node": "^22.14.0",
     "@types/react": "^18.3.20",
-    "@types/react-redux": "^7.1.34",
     "@types/react-dom": "^18.3.6",
+    "@types/react-redux": "^7.1.34",
     "@types/testing-library__react": "^10.0.1",
     "@typescript-eslint/eslint-plugin": "^8.24.1",
     "@typescript-eslint/parser": "^8.24.1",
@@ -63,7 +64,6 @@
   },
   "author": "Screenly, Inc.",
   "dependencies": {
-    "@reduxjs/toolkit": "^2.3.0",
     "bootstrap": "^5.3.3",
     "bootstrap-icons": "^1.11.3",
     "normalize-url": "^7.1.0",
