+++ src/assets/ts/features/asset/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './thunks';
+export * from './slice';
+export { default as assetReducer } from './slice';
