+++ src/assets/ts/store.ts
@@ -1,9 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
+import popupReducer from './features/popup-slice';
+import { assetReducer } from './features/asset';
 
-import popupReducer from '@/features/popup-slice';
-import assetReducer from '@/features/asset-slice';
-
-export const store = configureStore({
+const store = configureStore({
   reducer: {
     popup: popupReducer,
     asset: assetReducer,
@@ -12,3 +11,5 @@ export const store = configureStore({
 
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
+
+export { store };
