+++ src/assets/ts/types/store.ts
@@ -1,7 +1,10 @@
-import { AssetState } from '@/features/asset-slice';
+import { AssetState } from '@/features/asset';
 import { PopupState } from '@/features/popup-slice';
+import { store } from '@/store';
 
 export interface RootState {
   asset: AssetState;
   popup: PopupState;
 }
+
+export type AppDispatch = typeof store.dispatch;
